Refatoração do metodo remover da classe TopicosController

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -72,8 +72,12 @@ public class TopicosController {
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id){
-        topicoRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        Optional<Topico> optional = topicoRepository.findById(id);
+            if(optional.isPresent()){
+                topicoRepository.deleteById(id);
+                return ResponseEntity.ok().build();
+            }
+        return ResponseEntity.notFound().build();
     }
